+++ Solutions/Corvus.Json.ExtendedTypes/Corvus.Json.ExtendedTypes.csproj
@@ -27,11 +27,6 @@
     <PackageReference Include="System.Text.Json" Version="6.0.4" />
   </ItemGroup>
 
-
-  <ItemGroup>
-    <ProjectReference Include="..\Corvus.JsonHelpers\Corvus.JsonHelpers.csproj" />
-  </ItemGroup>
-
   <ItemGroup>
     <PackageReference Update="StyleCop.Analyzers" Version="1.2.0-beta.333" />
   </ItemGroup>
