+++ Solutions/Corvus.Json.ExtendedTypes/Corvus.Json/EncodedContentMediaTypeParseStatus.cs
