+++ Solutions/Corvus.Json.ExtendedTypes/Corvus.Json/Internal/JsonReaderHelper.cs
@@ -1,0 +1,158 @@
+﻿// <copyright file="JsonReaderHelper.cs" company="Endjin Limited">
+// Copyright (c) Endjin Limited. All rights reserved.
+// </copyright>
+// Derived from code:
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#pragma warning disable
+
+using System;
+using System.Buffers;
+using System.Buffers.Text;
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Text;
+
+namespace Corvus.Json
+{
+    internal static partial class JsonReaderHelper
+    {
+        public static ReadOnlySpan<byte> GetUnescapedSpan(ReadOnlySpan<byte> utf8Source, int idx)
+        {
+            // The escaped name is always >= than the unescaped, so it is safe to use escaped name for the buffer length.
+            int length = utf8Source.Length;
+            byte[]? pooledName = null;
+
+            Span<byte> utf8Unescaped = length <= JsonConstants.StackallocThreshold ?
+                stackalloc byte[length] :
+                (pooledName = ArrayPool<byte>.Shared.Rent(length));
+
+            Unescape(utf8Source, utf8Unescaped, idx, out int written);
+            Debug.Assert(written > 0);
+
+            ReadOnlySpan<byte> propertyName = utf8Unescaped.Slice(0, written).ToArray();
+            Debug.Assert(!propertyName.IsEmpty);
+
+            if (pooledName != null)
+            {
+                new Span<byte>(pooledName, 0, written).Clear();
+                ArrayPool<byte>.Shared.Return(pooledName);
+            }
+
+            return propertyName;
+        }
+
+        internal static void Unescape(ReadOnlySpan<byte> source, Span<byte> destination, int idx, out int written)
+        {
+            Debug.Assert(idx >= 0 && idx < source.Length);
+            Debug.Assert(source[idx] == JsonConstants.BackSlash);
+            Debug.Assert(destination.Length >= source.Length);
+
+            source.Slice(0, idx).CopyTo(destination);
+            written = idx;
+
+            for (; idx < source.Length; idx++)
+            {
+                byte currentByte = source[idx];
+                if (currentByte == JsonConstants.BackSlash)
+                {
+                    idx++;
+                    currentByte = source[idx];
+
+                    if (currentByte == JsonConstants.Quote)
+                    {
+                        destination[written++] = JsonConstants.Quote;
+                    }
+                    else if (currentByte == 'n')
+                    {
+                        destination[written++] = JsonConstants.LineFeed;
+                    }
+                    else if (currentByte == 'r')
+                    {
+                        destination[written++] = JsonConstants.CarriageReturn;
+                    }
+                    else if (currentByte == JsonConstants.BackSlash)
+                    {
+                        destination[written++] = JsonConstants.BackSlash;
+                    }
+                    else if (currentByte == JsonConstants.Slash)
+                    {
+                        destination[written++] = JsonConstants.Slash;
+                    }
+                    else if (currentByte == 't')
+                    {
+                        destination[written++] = JsonConstants.Tab;
+                    }
+                    else if (currentByte == 'b')
+                    {
+                        destination[written++] = JsonConstants.BackSpace;
+                    }
+                    else if (currentByte == 'f')
+                    {
+                        destination[written++] = JsonConstants.FormFeed;
+                    }
+                    else if (currentByte == 'u')
+                    {
+                        // The source is known to be valid JSON, and hence if we see a \u, it is guaranteed to have 4 hex digits following it
+                        // Otherwise, the Utf8JsonReader would have alreayd thrown an exception.
+                        Debug.Assert(source.Length >= idx + 5);
+
+                        bool result = Utf8Parser.TryParse(source.Slice(idx + 1, 4), out int scalar, out int bytesConsumed, 'x');
+                        Debug.Assert(result);
+                        Debug.Assert(bytesConsumed == 4);
+                        idx += bytesConsumed;     // The loop iteration will increment idx past the last hex digit
+
+                        if (JsonHelpers.IsInRangeInclusive((uint)scalar, JsonConstants.HighSurrogateStartValue, JsonConstants.LowSurrogateEndValue))
+                        {
+                            // The first hex value cannot be a low surrogate.
+                            if (scalar >= JsonConstants.LowSurrogateStartValue)
+                            {
+                                throw new InvalidOperationException($"Read Invalid UTF16: {scalar}");
+                            }
+
+                            Debug.Assert(JsonHelpers.IsInRangeInclusive((uint)scalar, JsonConstants.HighSurrogateStartValue, JsonConstants.HighSurrogateEndValue));
+
+                            idx += 3;   // Skip the last hex digit and the next \u
+
+                            // We must have a low surrogate following a high surrogate.
+                            if (source.Length < idx + 4 || source[idx - 2] != '\\' || source[idx - 1] != 'u')
+                            {
+                                throw new InvalidOperationException("Read Invalid UTF16");
+                            }
+
+                            // The source is known to be valid JSON, and hence if we see a \u, it is guaranteed to have 4 hex digits following it
+                            // Otherwise, the Utf8JsonReader would have alreayd thrown an exception.
+                            result = Utf8Parser.TryParse(source.Slice(idx, 4), out int lowSurrogate, out bytesConsumed, 'x');
+                            Debug.Assert(result);
+                            Debug.Assert(bytesConsumed == 4);
+
+                            // If the first hex value is a high surrogate, the next one must be a low surrogate.
+                            if (!JsonHelpers.IsInRangeInclusive((uint)lowSurrogate, JsonConstants.LowSurrogateStartValue, JsonConstants.LowSurrogateEndValue))
+                            {
+                                throw new InvalidOperationException($"Read Invalid UTF16: {lowSurrogate}");
+                            }
+
+                            idx += bytesConsumed - 1;  // The loop iteration will increment idx past the last hex digit
+
+                            // To find the unicode scalar:
+                            // (0x400 * (High surrogate - 0xD800)) + Low surrogate - 0xDC00 + 0x10000
+                            scalar = (JsonConstants.BitShiftBy10 * (scalar - JsonConstants.HighSurrogateStartValue))
+                                + (lowSurrogate - JsonConstants.LowSurrogateStartValue)
+                                + JsonConstants.UnicodePlane01StartValue;
+                        }
+
+                        var rune = new Rune(scalar);
+                        int bytesWritten = rune.EncodeToUtf8(destination.Slice(written));
+                        Debug.Assert(bytesWritten <= 4);
+                        written += bytesWritten;
+                    }
+                }
+                else
+                {
+                    destination[written++] = currentByte;
+                }
+            }
+        }
+    }
+}
