--- Solutions/Corvus.JsonHelpers/Corvus.JsonHelpers.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-	<Import Project="$(EndjinProjectPropsPath)" Condition="$(EndjinProjectPropsPath) != ''" />
-	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-	</PropertyGroup>
-
-	<PropertyGroup>
-		<PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-		<Description>Utility types and methods for working with JSON</Description>
-    </PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Corvus.Extensions" Version="1.1.4" />
-		<PackageReference Include="Endjin.RecommendedPractices" Version="1.2.0">
-			<PrivateAssets>all</PrivateAssets>
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-		</PackageReference>
-		<PackageReference Include="Microsoft.Extensions.Http" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.ObjectPool" Version="6.0.5" />
-		<PackageReference Include="NodaTime" Version="3.1.0" />
-		<PackageReference Include="System.Buffers" Version="4.5.1" />
-		<PackageReference Include="System.Collections.Immutable" Version="6.0.0" />
-		<PackageReference Include="System.Text.Json" Version="6.0.4" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Update="StyleCop.Analyzers" Version="1.2.0-beta.333" />
-	</ItemGroup>
-
-</Project>
