+++ Solutions/Corvus.JsonSchema.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Json.JsonSchema", "C
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Json.ExtendedTypes", "Corvus.Json.ExtendedTypes\Corvus.Json.ExtendedTypes.csproj", "{07C49DF8-CE8E-46F9-8DA4-F0B8DEFF1D5A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.JsonHelpers", "Corvus.JsonHelpers\Corvus.JsonHelpers.csproj", "{2B019910-FE10-4B86-A225-F34C6305E1F3}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Json.Walker", "Corvus.Json.Walker\Corvus.Json.Walker.csproj", "{378ED2EF-C859-4313-943C-F3B2A8F76C59}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Code Generation", "Code Generation", "{87E7B0E0-1D4F-4764-80F3-7DB33A014E88}"
@@ -47,10 +45,6 @@ Global
 		{07C49DF8-CE8E-46F9-8DA4-F0B8DEFF1D5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07C49DF8-CE8E-46F9-8DA4-F0B8DEFF1D5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07C49DF8-CE8E-46F9-8DA4-F0B8DEFF1D5A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2B019910-FE10-4B86-A225-F34C6305E1F3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B019910-FE10-4B86-A225-F34C6305E1F3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B019910-FE10-4B86-A225-F34C6305E1F3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B019910-FE10-4B86-A225-F34C6305E1F3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{378ED2EF-C859-4313-943C-F3B2A8F76C59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{378ED2EF-C859-4313-943C-F3B2A8F76C59}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{378ED2EF-C859-4313-943C-F3B2A8F76C59}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -86,7 +80,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{83A532B0-4096-4EEF-A12B-63C739155D62} = {8F54F97B-19FF-46DB-8EE7-47E3079D30A7}
 		{07C49DF8-CE8E-46F9-8DA4-F0B8DEFF1D5A} = {8F54F97B-19FF-46DB-8EE7-47E3079D30A7}
-		{2B019910-FE10-4B86-A225-F34C6305E1F3} = {8F54F97B-19FF-46DB-8EE7-47E3079D30A7}
 		{378ED2EF-C859-4313-943C-F3B2A8F76C59} = {87E7B0E0-1D4F-4764-80F3-7DB33A014E88}
 		{CDA2599D-DDF1-432E-8A38-C067DE96EF04} = {87E7B0E0-1D4F-4764-80F3-7DB33A014E88}
 		{71DBE0A7-C346-4513-8F86-649410018A99} = {97A39EA1-C74F-4B7D-99CB-9FD202A155E7}
